refactor: FE User Card now mui5 styled
